requirements.txt
asgiref==3.5.2
Django==4.0.5
django-ckeditor==6.4.2
django-ckeditor-updated==4.4.4
django-js-asset==2.0.0
djangorestframework==3.13.1
mysqlclient==2.1.0
Pillow==9.1.1
python-slugify==6.1.2
pytz==2022.1
sqlparse==0.4.2
text-unidecode==1.3
